data-enrichment-system/requirements.txt
streamlit==1.28.0
pandas==2.1.0
openpyxl==3.1.2
scrapy==2.11.0
requests==2.31.0
openai==0.28.1
python-dotenv==1.0.0
validators==0.22.0
beautifulsoup4==4.12.2
lxml==4.9.3
fake-useragent==1.4.0
selenium==4.15.0
webdriver-manager==4.0.1
redis==5.0.0
celery==5.3.1
asyncio==3.4.3
aiohttp==3.8.5
tenacity==8.2.3